remove the network package as well
