+++ cmd/benchmark.go
@@ -12,12 +12,12 @@ import (
 	"github.com/quorumcontrol/chaintree/chaintree"
 	"github.com/quorumcontrol/chaintree/nodestore"
 	"github.com/quorumcontrol/chaintree/safewrap"
+	"github.com/quorumcontrol/storage"
 	"github.com/quorumcontrol/tupelo/bls"
 	"github.com/quorumcontrol/tupelo/consensus"
 	"github.com/quorumcontrol/tupelo/gossip2"
 	"github.com/quorumcontrol/tupelo/gossip2client"
-	"github.com/quorumcontrol/tupelo/network"
-	"github.com/quorumcontrol/storage"
+	"github.com/quorumcontrol/tupelo/p2p"
 	"github.com/spf13/cobra"
 )
 
@@ -172,7 +172,7 @@ var benchmark = &cobra.Command{
 			testNetMembers := bootstrapMembers(bootstrapPublicKeys)
 			group.CreateGenesisState(group.RoundAt(time.Now()), testNetMembers...)
 		}
-		client := gossip2client.NewGossipClient(group, network.BootstrapNodes())
+		client := gossip2client.NewGossipClient(group, p2p.BootstrapNodes())
 
 		time.Sleep(5 * time.Second)
